pyproject.toml
[project]
name = "rafam-workflow"
version = "0.2.0"
description = "Base app para gestionar flujo de suministros y presupuesto"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlmodel>=0.0.22",
  "pydantic-settings>=2.4.0",
  "python-multipart>=0.0.9",
  "openpyxl>=3.1.5",
  "pypdf>=4.3.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.27.0"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
